requirements.dag.txt
apache-airflow[crypto,celery,postgres,hive,jdbc,ssh]==1.10.10
dask[complete]==2.15.0
distributed==2.15.2
